+++ internal/controller/operator/factory/reconcile/hpa.go
@@ -21,6 +21,7 @@ func HPA(ctx context.Context, rclient client.Client, newObj, prevObj *v2.Horizon
 	if prevObj != nil {
 		prevMeta = &prevObj.ObjectMeta
 	}
+	removeFinalizer := true
 	return retryOnConflict(func() error {
 		var existingObj v2.HorizontalPodAutoscaler
 		if err := rclient.Get(ctx, nsn, &existingObj); err != nil {
@@ -30,10 +31,10 @@ func HPA(ctx context.Context, rclient client.Client, newObj, prevObj *v2.Horizon
 			}
 			return fmt.Errorf("cannot get HPA=%s: %w", nsn.String(), err)
 		}
-		if err := collectGarbage(ctx, rclient, &existingObj); err != nil {
+		if err := collectGarbage(ctx, rclient, &existingObj, removeFinalizer); err != nil {
 			return err
 		}
-		metaChanged, err := mergeMeta(&existingObj, newObj, prevMeta, owner, true)
+		metaChanged, err := mergeMeta(&existingObj, newObj, prevMeta, owner, removeFinalizer, false)
 		if err != nil {
 			return err
 		}