+++ internal/controller/operator/factory/reconcile/service.go
@@ -32,6 +32,7 @@ func reconcileService(ctx context.Context, rclient client.Client, newObj, prevOb
 	// helper for proper service deletion.
 	nsn := types.NamespacedName{Name: newObj.Name, Namespace: newObj.Namespace}
 	var existingObj corev1.Service
+	removeFinalizer := true
 	if err := rclient.Get(ctx, nsn, &existingObj); err != nil {
 		if k8serrors.IsNotFound(err) {
 			logger.WithContext(ctx).Info(fmt.Sprintf("creating new Service=%s", nsn.String()))
@@ -43,7 +44,7 @@ func reconcileService(ctx context.Context, rclient client.Client, newObj, prevOb
 		}
 		return fmt.Errorf("cannot get service for existing Service=%s: %w", nsn.String(), err)
 	}
-	if err := collectGarbage(ctx, rclient, &existingObj); err != nil {
+	if err := collectGarbage(ctx, rclient, &existingObj, removeFinalizer); err != nil {
 		return err
 	}
 	var prevMeta *metav1.ObjectMeta
@@ -114,7 +115,7 @@ func reconcileService(ctx context.Context, rclient client.Client, newObj, prevOb
 	}
 
 	rclient.Scheme().Default(newObj)
-	metaChanged, err := mergeMeta(&existingObj, newObj, prevMeta, owner, true)
+	metaChanged, err := mergeMeta(&existingObj, newObj, prevMeta, owner, removeFinalizer, false)
 	if err != nil {
 		return err
 	}