+++ internal/controller/operator/factory/reconcile/service_account.go
@@ -20,6 +20,7 @@ func ServiceAccount(ctx context.Context, rclient client.Client, newObj, prevObj 
 	if prevObj != nil {
 		prevMeta = &prevObj.ObjectMeta
 	}
+	removeFinalizer := true
 	return retryOnConflict(func() error {
 		var existingObj corev1.ServiceAccount
 		if err := rclient.Get(ctx, nsn, &existingObj); err != nil {
@@ -29,10 +30,10 @@ func ServiceAccount(ctx context.Context, rclient client.Client, newObj, prevObj 
 			}
 			return fmt.Errorf("cannot get ServiceAccount=%s: %w", nsn.String(), err)
 		}
-		if err := collectGarbage(ctx, rclient, &existingObj); err != nil {
+		if err := collectGarbage(ctx, rclient, &existingObj, removeFinalizer); err != nil {
 			return err
 		}
-		metaChanged, err := mergeMeta(&existingObj, newObj, prevMeta, owner, true)
+		metaChanged, err := mergeMeta(&existingObj, newObj, prevMeta, owner, removeFinalizer, false)
 		if err != nil {
 			return err
 		}