+++ internal/controller/operator/factory/reconcile/vmservicescrape_test.go
@@ -46,7 +46,7 @@ func TestVMServiceScrape(t *testing.T) {
 		t.Helper()
 		ctx := context.Background()
 		cl := k8stools.GetTestClientWithActionsAndObjects(o.predefinedObjects)
-		assert.NoError(t, VMServiceScrape(ctx, cl, o.new, o.prev, nil))
+		assert.NoError(t, VMServiceScrape(ctx, cl, o.new, o.prev, nil, false))
 		assert.Equal(t, o.actions, cl.Actions)
 	}