+++ internal/controller/operator/factory/vmanomaly/statefulset.go
@@ -56,7 +56,7 @@ func CreateOrUpdate(ctx context.Context, cr *vmv1.VMAnomaly, rclient client.Clie
 	if !ptr.Deref(cr.Spec.DisableSelfServiceScrape, false) {
 		svs := buildScrape(cr)
 		prevSvs := buildScrape(prevCR)
-		if err := reconcile.VMPodScrape(ctx, rclient, svs, prevSvs, &owner); err != nil {
+		if err := reconcile.VMPodScrape(ctx, rclient, svs, prevSvs, &owner, false); err != nil {
 			return err
 		}
 	}